Update Laravel:
Following items have been added:

Trait HttpResponses, contains a easy way to send back success and error messages

Controller AuthController, This allows the creation of users and allows users to log in (log out not yet implemented
Controller AchievementController, Created but no features as of now
Controller ExerciseController, Crud api for exercises

Kernal and Authenticate have been changed, kernal allowed sanctum to be uncluded, for authentication and Authenticate so there is no error for not having a login page

Added models for Users Exercises PersonalAccessTokens and Achievements. There is not yet a relation between Exercise and Achievement Within these. Need to be added

Renamed Migrations to now work on every pc, this has to do with the relation table needing exercise to be created before achievements. This was not a problem on every pc, but this is the way to do it.

Updated routes to include a public and protected section, Added all endpoints for Exercise